Fixed numeric types should map to decimal.

+++ src/responses/types/row_type.rs
@@ -53,7 +53,7 @@ impl RowType {
                     }
                 }
                 else {
-                    ValueType::Float
+                    ValueType::Decimal
                 }
             }
             "real" => ValueType::Float,
